Split python into two files, un/recomment stuff
