+++ .vscode/settings.json
@@ -4,4 +4,5 @@
     "--background-index",
     "--compile-commands-dir=build_debug"
   ],
+  "cmake.configureOnEdit": false,
 }
